refactor: move migration to own module
